Float32ized

+++ src/Utils.jl
@@ -18,7 +18,7 @@ function common_parse_commandline()
         "--pseudocount", "-p"
             help = "log10(exp + pseudocount)"
             arg_type = Float32
-            default = 1.0
+            default = Float32(1.0)
         "--rowmeanlist", "-m"
             help = "mean vector of each row"
             arg_type = String
@@ -41,7 +41,7 @@ function common_parse_commandline()
         "--stepsize", "-s"
             help = "stepsize of PCA"
             arg_type = Float32
-            default = 0.1
+            default = Float32(0.1)
         "--numepoch", "-e"
             help = "numepoch of PCA"
             arg_type = Int64
@@ -53,11 +53,11 @@ function common_parse_commandline()
         "-g"
             help = "Ratio of non-SGD gradient"
             arg_type = Float32
-            default = 0.9
+            default = Float32(0.9)
         "--epsilon"
             help = "a small number for avoiding zero division"
             arg_type = Float32
-            default = 1.0e-8
+            default = Float32(1.0e-8)
         "--logdir", "-l"
             help = "saving log directory"
             default = nothing
