+++ src/rsgd.jl
@@ -1,5 +1,5 @@
 """
-    rsgd(;input::String="", outdir=nothing, logscale::Bool=true, pseudocount::Float32=1, rowmeanlist::String="", colsumlist::String="", masklist::String="", dim::Int64=3, stepsize::Float32=0.1, numepoch::Int64=5, scheduling::String="robbins-monro", g::Float32=0.9, epsilon::Float32=1.0e-8, logdir=nothing)
+    rsgd(;input::String="", outdir=nothing, logscale::Bool=true, pseudocount::Float32=Float32(1), rowmeanlist::String="", colsumlist::String="", masklist::String="", dim::Int64=3, stepsize::Float32=Float32(0.1), numepoch::Int64=5, scheduling::String="robbins-monro", g::Float32=Float32(0.9), epsilon::Float32=Float32(1.0e-8), logdir=nothing)
 
 Online PCA solved by Riemannian stochastic gradient descent method.
 
@@ -30,7 +30,7 @@ Reference
 ---------
 - RSGD-PCA : [Silvere Bonnabel, 2013](https://arxiv.org/abs/1111.5280)
 """
-function rsgd(;input::String="", outdir=nothing, logscale::Bool=true, pseudocount::Float32=1, rowmeanlist::String="", colsumlist::String="", masklist::String="", dim::Int64=3, stepsize::Float32=0.1, numepoch::Int64=5, scheduling::String="robbins-monro", g::Float32=0.9, epsilon::Float32=1.0e-8, logdir=nothing)
+function rsgd(;input::String="", outdir=nothing, logscale::Bool=true, pseudocount::Float32=Float32(1), rowmeanlist::String="", colsumlist::String="", masklist::String="", dim::Int64=3, stepsize::Float32=Float32(0.1), numepoch::Int64=5, scheduling::String="robbins-monro", g::Float32=Float32(0.9), epsilon::Float32=Float32(1.0e-8), logdir=nothing)
     # Initialization
     const N::Int64, M::Int64 = init(input) # No.gene, No.cell
     W = zeros(Float32, M, dim) # Eigen vectors
